backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication (Supabase JWT verification)
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Background tasks
arq>=0.26.0
redis>=5.0.0

# External services
httpx>=0.27.0
twilio>=9.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
